requirements_1.txt
aiomysql==0.2.0
alembic==1.15.2
altair==4.2.2
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1733250440834/work
async-lru==2.0.5
attrs==25.3.0
avro==1.11.3
avro-python3==1.10.2
babel==2.17.0
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
boto3==1.38.13
botocore==1.38.13
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1733502965406/work
confluent-kafka==2.6.1
contourpy==1.3.2
cryptography==44.0.3
cycler==0.12.1
databricks-sdk==0.52.0
debugpy @ file:///D:/bld/debugpy_1680755734210/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1740384970518/work
defusedxml==0.7.1
Deprecated==1.2.18
docker==7.1.0
entrypoints==0.4
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1733208806608/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1745502089858/work
fastapi==0.115.12
fastavro==1.8.4
fastjsonschema==2.21.1
Flask==3.1.0
fonttools==4.57.0
fqdn==1.5.1
fsspec==2025.3.2
furl==2.1.4
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.40.1
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
great-expectations==0.18.12
greenlet==3.2.2
grpcio==1.71.0
h11==0.16.0
hopsworks==4.2.2
hopsworks-aiomysql==0.2.1
hsfs==3.7.9
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1737420181517/work
ipykernel @ file:///D:/bld/ipykernel_1719845595208/work
ipython @ file:///D:/bld/bld/rattler-build_ipython_1745672185/work
ipython_pygments_lexers @ file:///home/conda/feedstock_root/build_artifacts/ipython_pygments_lexers_1737123620466/work
ipywidgets==8.1.7
isoduration==20.11.0
itsdangerous==2.2.0
javaobj-py3==0.4.4
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1733300866624/work
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.0
json5==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1733440914442/work
jupyter_core @ file:///D:/bld/jupyter_core_1727163532151/work
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kiwisolver==1.4.8
lightgbm==4.6.0
makefun==1.16.0
Mako==1.3.10
Markdown==3.8
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1733416936468/work
mistune==3.1.3
mlflow==2.22.0
mlflow-skinny==2.22.0
mock==5.2.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1733325553580/work
notebook==7.4.2
notebook_shim==0.2.4
numpy==1.26.4
opensearch-py==2.4.2
opentelemetry-api==1.33.0
opentelemetry-sdk==1.33.0
opentelemetry-semantic-conventions==0.54b0
orderedmultidict==1.0.1
overrides==7.7.0
packaging==24.2
pandas==2.1.4
pandocfilters==1.5.1
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1733271261340/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1733327343728/work
pillow==11.2.1
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_platformdirs_1746710438/work
prometheus_client==0.21.1
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1744724089886/work
protobuf==4.25.7
psutil @ file:///D:/bld/psutil_1681775154857/work
pure_eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1733569405015/work
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodomex==3.22.0
pydantic==2.11.4
pydantic_core==2.33.2
pydeck==0.9.1
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1736243443484/work
pyhumps==1.6.1
pyjks==20.0.0
PyMySQL==1.1.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==3.3.0
pytz==2025.2
pywin32==304
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq @ file:///D:/bld/pyzmq_1679317063994/work
referencing==0.36.2
requests==2.32.3
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.24.0
rsa==4.9.1
ruamel.yaml==0.17.17
s3transfer==0.12.0
scikit-learn==1.6.1
scipy==1.15.3
Send2Trash==1.8.3
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==1.4.48
sqlparse==0.5.3
stack_data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1733569443808/work
starlette==0.46.2
streamlit==1.45.0
tenacity==9.1.2
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
toml==0.10.2
toolz==1.0.0
tornado @ file:///D:/bld/tornado_1681817639987/work
tqdm==4.67.1
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1733367359838/work
twofish @ file:///C:/ci_311/twofish_1676495280622/work
types-python-dateutil==2.9.0.20241206
typing-inspection==0.4.0
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_typing_extensions_1744302253/work
tzdata==2025.2
tzlocal==5.3.1
uri-template==1.3.0
urllib3==2.4.0
uvicorn==0.34.2
waitress==3.0.2
watchdog==6.0.0
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1733231326287/work
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
widgetsnbextension==4.0.14
wrapt==1.17.2
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1732827521216/work
